Extract console module, write unit tests
This expands the debugger console to add most of the standard features
from C's printf. The functionality is tested with a unit test. A unit
test framework is added which runs on the Nintendo 64. This is all one
change because it is tied so closely together: the test framework relies
on the console library in order to display test output on the N64.

GitOrigin-RevId: 8c35709f4c73771d0a3ed4dfad992ee4f9dc6f5e